Introduce consumer blueprint and migrate endpoints
