+++ app/routes/__init__.py
@@ -3,7 +3,7 @@ from .user import user_bp
 from .vendor import vendor_bp
 from .shop import shop_bp
 from .item import item_bp
-from .cart import cart_bp
+from .consumer import consumer_bp
 from .wallet import wallet_bp
 from .order import order_bp
 from .admin import admin_bp
@@ -19,7 +19,7 @@ __all__ = [
     'vendor_bp',
     'shop_bp',
     'item_bp',
-    'cart_bp',
+    'consumer_bp',
     'wallet_bp',
     'order_bp',
     'admin_bp',
