+++ app/routes/shop.py
@@ -200,89 +200,3 @@ def toggle_shop_status():
 
     return jsonify({"status": "success", "message": f"Shop marked as {action}"}), 200
 
-
-# --- Shop searching by customer ---
-
-# List all shops in the user's society and city
-@auth_required
-@role_required(["consumer"])
-def list_shops():
-    user = request.user
-    city, society = user.city, user.society
-
-    # Optional query filters
-    status = request.args.get("status")       # e.g. "open" or "closed"
-    shop_type = request.args.get("type")      # e.g. "grocery"
-    tags = request.args.getlist("tag")        # e.g. ?tag=organic&tag=dairy
-
-    query = Shop.query.filter_by(city=city, society=society)
-
-    if status == "open":
-        query = query.filter_by(is_open=True)
-    elif status == "closed":
-        query = query.filter_by(is_open=False)
-
-    if shop_type:
-        query = query.filter(Shop.shop_type.ilike(f"%{shop_type}%"))
-
-    if tags:
-        # assume category_tags stored as comma‐separated or JSON array
-        for t in tags:
-            query = query.filter(Shop.category_tags.ilike(f"%{t}%"))
-
-    shops = query.all()
-    result = []
-    for s in shops:
-        result.append({
-            "id": s.id,
-            "shop_name": s.shop_name,
-            "shop_type": s.shop_type,
-            "description": s.description,
-            "is_open": s.is_open,
-            "delivers": s.delivers,
-            "appointment_only": s.appointment_only,
-            "category_tags": s.category_tags,
-            "logo_url": s.logo_url
-        })
-
-    return jsonify({"status": "success", "shops": result}), 200
-
-# Search shops by name or type
-@shop_bp.route("/shops/search", methods=["GET"])
-@auth_required
-@role_required(["consumer"])
-def search_shops():
-    user = request.user
-    city = user.city
-    society = user.society
-
-    query_param = request.args.get("q", "").lower().strip()
-
-    if not query_param:
-        return error("Missing search query 'q'", status=400)
-
-    results = Shop.query.filter(
-        Shop.city == city,
-        Shop.society == society,
-        Shop.is_open == True,
-        db.or_(
-            Shop.shop_name.ilike(f"%{query_param}%"),
-            Shop.shop_type.ilike(f"%{query_param}%")
-        )
-    ).all()
-
-    shop_list = []
-    for shop in results:
-        shop_list.append({
-            "id": shop.id,
-            "shop_name": shop.shop_name,
-            "shop_type": shop.shop_type,
-            "description": shop.description,
-            "is_open": shop.is_open,
-            "delivers": shop.delivers,
-            "appointment_only": shop.appointment_only,
-            "category_tags": shop.category_tags,
-            "logo_url": shop.logo_url
-        })
-
-    return jsonify({"status": "success", "shops": shop_list}), 200
