+++ app/routes/wallet.py
@@ -5,11 +5,8 @@ from flask import request, jsonify
 from app.version import API_PREFIX
 from models import db
 from models.wallet import (
-    ConsumerWallet,
-    WalletTransaction,
     VendorWallet,
     VendorWalletTransaction,
-
 )
 from models.vendor import VendorPayoutBank
 wallet_bp = Blueprint("wallet", __name__, url_prefix=API_PREFIX)
@@ -20,124 +17,11 @@ import logging
 from app.utils import internal_error_response
 from app.utils import error, transactional
 from app.services.wallet_ops import (
-    adjust_consumer_balance,
     adjust_vendor_balance,
     InsufficientFunds,
 )
 
 
-# ------------------- Get or Create Consumer Wallet -------------------
-@wallet_bp.route("/wallet", methods=["GET"])
-@auth_required
-@role_required(["consumer"])
-def get_or_create_wallet():
-    """Return the consumer wallet for the authenticated user, creating one if needed."""
-    user = request.user
-    wallet = ConsumerWallet.query.filter_by(user_phone=user.phone).first()
-    if not wallet:
-        wallet = ConsumerWallet(user_phone=user.phone, balance=Decimal("0.00"))
-        try:
-            with transactional("Failed to create wallet"):
-                db.session.add(wallet)
-        except Exception:
-            return internal_error_response()
-    return jsonify({"status": "success", "balance": float(wallet.balance)}), 200
-
-
-# ------------------- Consumer Wallet Transactions -------------------
-@wallet_bp.route("/wallet/history", methods=["GET"])
-@auth_required
-@role_required(["consumer"])
-def wallet_transaction_history():
-    txns = WalletTransaction.query.filter_by(user_phone=request.phone)\
-        .order_by(WalletTransaction.created_at.desc()).limit(50).all()
-    result = [txn.to_dict() for txn in txns]
-    return jsonify({"status": "success", "transactions": result}), 200
-
-
-@wallet_bp.route("/wallet/load", methods=["POST"])
-# ------------------- Load Consumer Wallet -------------------
-@auth_required
-@role_required(["consumer"])
-def load_wallet():
-    """Add funds to the authenticated consumer's wallet."""
-    user = request.user
-    data = request.get_json()
-    try:
-        amount = Decimal(str(data.get("amount", "0")))
-        if amount <= 0:
-            return error("Invalid amount", status=400)
-
-        with transactional("Failed to load wallet"):
-            new_bal = adjust_consumer_balance(
-                user.phone,
-                amount,
-                reference=data.get("reference", "manual-load"),
-                type="recharge",
-                source="api",
-            )
-        return jsonify({"status": "success", "balance": float(new_bal)}), 200
-    except InsufficientFunds as e:
-        return error(str(e), status=400)
-    except Exception as e:
-        logging.error("Failed to load wallet: %s", e, exc_info=True)
-        return internal_error_response()
-
-@wallet_bp.route("/wallet/debit", methods=["POST"])
-
-# ------------------- Debit Consumer Wallet -------------------
-@auth_required
-@role_required(["consumer"])
-def debit_wallet():
-    """Deduct funds from the authenticated consumer's wallet."""
-    user = request.user
-    data = request.get_json()
-    try:
-        amount = Decimal(str(data.get("amount", "0")))
-        reference = data.get("reference", "manual-debit")
-
-        with transactional("Failed to debit wallet"):
-            new_bal = adjust_consumer_balance(
-                user.phone,
-                -amount,
-                reference=reference,
-                type="debit",
-                source="api",
-            )
-        return jsonify({"status": "success", "balance": float(new_bal)}), 200
-    except InsufficientFunds as e:
-        return error(str(e), status=400)
-    except Exception as e:
-        logging.error("Failed to debit wallet: %s", e, exc_info=True)
-        return internal_error_response()
-
-
-# ------------------- Refund to Consumer Wallet -------------------
-@wallet_bp.route("/wallet/refund", methods=["POST"])
-@auth_required
-@role_required(["consumer"])
-def refund_wallet():
-    """Refund an amount back to the consumer's wallet, creating it if needed."""
-    user = request.user
-    data = request.get_json()
-    try:
-        amount = Decimal(str(data.get("amount", "0")))
-        reference = data.get("reference", "manual-refund")
-
-        with transactional("Failed to refund wallet"):
-            new_bal = adjust_consumer_balance(
-                user.phone,
-                amount,
-                reference=reference,
-                type="refund",
-                source="api",
-            )
-        return jsonify({"status": "success", "balance": float(new_bal)}), 200
-    except InsufficientFunds as e:
-        return error(str(e), status=400)
-    except Exception as e:
-        logging.error("Failed to refund wallet: %s", e, exc_info=True)
-        return internal_error_response()
 
 
 # ------------------- Get or Create Vendor Wallet -------------------
